[FIX] 공간 방문 api 라우터 수정 #43
