+++ src/main/java/indipage/org/indipage/api/space/service/SpaceService.java
@@ -4,19 +4,26 @@ import indipage.org.indipage.api.space.controller.dto.response.BookRecommendatio
 import indipage.org.indipage.api.space.controller.dto.response.FollowSpaceRelationResponseDto;
 import indipage.org.indipage.api.space.controller.dto.response.SpaceDto;
 import indipage.org.indipage.api.user.service.UserService;
-import indipage.org.indipage.domain.*;
+import indipage.org.indipage.domain.FollowSpaceRelationRepository;
+import indipage.org.indipage.domain.InviteSpaceRelationRepository;
 import indipage.org.indipage.domain.Relation.BookRecommendationRelation;
 import indipage.org.indipage.domain.Relation.FollowSpaceRelation;
 import indipage.org.indipage.domain.Relation.FollowSpaceRelationId;
+import indipage.org.indipage.domain.Relation.InviteSpaceRelation;
+import indipage.org.indipage.domain.Relation.InviteSpaceRelationId;
 import indipage.org.indipage.domain.Relation.SpaceTagRelation;
+import indipage.org.indipage.domain.Space;
+import indipage.org.indipage.domain.SpaceRepository;
+import indipage.org.indipage.domain.Tag;
+import indipage.org.indipage.domain.User;
 import indipage.org.indipage.exception.Error;
 import indipage.org.indipage.exception.model.NotFoundException;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
+import javax.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
@@ -24,6 +31,7 @@ public class SpaceService {
 
     private final SpaceRepository spaceRepository;
     private final FollowSpaceRelationRepository followSpaceRelationRepository;
+    private final InviteSpaceRelationRepository inviteSpaceRelationRepository;
     private final UserService userService;
 
     public SpaceDto readSpace(final Long spaceId) {
@@ -47,6 +55,22 @@ public class SpaceService {
         }
 
         return responseDtos;
+    }
+
+    @Transactional(rollbackOn = Exception.class)
+    public void visit(final Long userId, final Long spaceId) {
+
+        User user = userService.findUser(userId);
+        Space space = findSpace(spaceId);
+
+        InviteSpaceRelation relation = inviteSpaceRelationRepository.findByInviteSpaceRelationId(
+                InviteSpaceRelationId.newInstance(user, space)).orElseThrow(
+                () -> new NotFoundException(Error.NOT_FOUND_TICKET_RECEIVE_EXCEPTION,
+                        Error.NOT_FOUND_TICKET_RECEIVE_EXCEPTION.getMessage()));
+
+        // 방문하기
+        relation.visit();
+        inviteSpaceRelationRepository.save(relation);
     }
 
     public FollowSpaceRelationResponseDto readFollowSpace(final Long userId, final Long spaceId) {
